Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2
scikit-learn==1.3.2
pytest==7.4.3
httpx==0.25.2
spacy==3.7.2
aiohttp==3.9.1